change: move unit tests
